% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ainv}
\alias{ainv}
\title{Pedigree data in ASREML format.}
\format{
A Large matrix with 31689 elements(By asreml::ainverse):
}
\usage{
ainv
}
\description{
This data can be generated by asreml::ainverse(PED).
}
\keyword{datasets}
