% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueEstimation1.0.R
\name{blueEstimation}
\alias{blueEstimation}
\title{BLUE value estimation}
\usage{
blueEstimation(datdf, repct, trait)
}
\arguments{
\item{datdf}{a data.frame}

\item{repct}{a number}

\item{trait}{trait name}
}
\value{
A list
}
\description{
BLUE value estimation
}
\examples{
\donttest{
names(phdemo)
dat<-phdemo$SU
dat <- dat[!is.na(dat$Range) & !is.na(dat$Pass),]
dat <- dat[!is.na(dat$Name),]
dat <- dat[!is.na(dat$Location),]
dat <- dat[!is.na(dat$Field),]
dat <- dat[!is.na(dat$Type),]
dat<-dat[dat$Location==unique(dat$Location)[1],]
dat<-dat[dat$Field==unique(dat$Field)[1],]
datdf<-dat[dat$Type==unique(dat$Type)[1],]

print(table(datdf$Range,datdf$Pass))

trait<-"YLD"
names(datdf)[names(datdf)=="YLD"]<-"trait"

repct<-replicationCheck(df=datdf)
repct
datdf<-missingAdding(w=datdf)
datdf <- datdf[order(datdf$Range,datdf$Pass),]
print(table(datdf$Range,datdf$Pass))
library(asreml)
blueout<-blueEstimation(datdf,repct,trait)
}
}
