% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{blueEstimationY}
\alias{blueEstimationY}
\title{Spatial correction (BLUE analysis)}
\usage{
blueEstimationY(
  datdf,
  trait,
  na.max = 0.9,
  fix.factor = "all",
  min.valid.range = 2,
  min.valid.pass = 2,
  verbose = TRUE
)
}
\description{
:
}
