% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{blupEstimation}
\alias{blupEstimation}
\title{BLUP analysis}
\usage{
blupEstimation(
  datdf,
  trait,
  na.max = 0.9,
  fix.factor = "all",
  min.valid.range = 2,
  min.valid.pass = 2,
  verbose = TRUE
)
}
\description{
The function is used for preliminary filtering of raw data, including:
}
