% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropEval-package.R
\docType{package}
\name{cropEval-package}
\alias{cropEval}
\alias{cropEval-package}
\title{cropEval: Qingyuan Genetic Evaluation software(Breeders aid decision making tool)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
