% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_checking.R
\name{dataOutliers.for}
\alias{dataOutliers.for}
\title{Identify and process outliers in the data for multiple traits}
\usage{
dataOutliers.for(data, traits, model)
}
\arguments{
\item{data}{The input data as a data.frame.}

\item{traits}{A character vector containing the names of the traits to be processed.}

\item{model}{The method to be used for identifying outliers, either '3SD' or 'boxplot'.}
}
\value{
A list containing processed data for each trait separately.
}
\description{
This function identifies and processes outliers in the data for multiple traits.
}
\examples{
w <- phdemo$SpML
yld <- dataOutliers.for(data=w,traits =c('MST',"YLD","PHT","EHT"),model='3SD')

}
