% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_arrange.R
\name{del_result}
\alias{del_result}
\title{Collation of genetic evaluation results}
\usage{
del_result(
  popmean,
  region,
  traits,
  ctrlnsample,
  ctrlssample,
  adv,
  hybobs,
  blupebv,
  Aov
)
}
\arguments{
\item{popmean}{Population average}

\item{region}{Maize varieties}

\item{traits}{phenotypic traits}

\item{ctrlnsample}{Reference male parent}

\item{ctrlssample}{Reference female parent}

\item{adv}{advanced individuals in 2022}

\item{hybobs}{No. of hybrid observations}

\item{blupebv}{breeding value}

\item{Aov}{Variance test table}
}
\value{
outebv A list of data frames
}
\description{
Collation of genetic evaluation results
}
\examples{
\donttest{
rm(list = ls())
library(readxl)
library(writexl)
library(biosim)
library(tidyverse)
library(reshape2)
library(stringi)
library(ggplot2)
library(ggpmisc)
region = 'SU'
traits <- c('YLD','MST','TWT','PHT','EHT','PMD','LDG')
ctrlnsample = 'Chang72'  # Reference male parent
ctrlssample = 'Zheng58'  # Reference female parent
del_result_list <- del_result(popmean, region, traits, ctrlnsample, ctrlssample, adv, hybobs, blupebv, Aov)
print(names(del_result_list))
}

}
