% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPbreedingValues_Estimation.R
\name{extracting}
\alias{extracting}
\title{Extracting some important statistics}
\usage{
extracting(asr, var, individual, term)
}
\arguments{
\item{asr}{Established model.}

\item{var}{Variance component corresponding to factor.}

\item{individual}{Samples corresponding to factors.}

\item{term}{The name of each factor displayed in the model results.}

\item{gxe}{Identifier.}
}
\value{
A data frame with important statistics.
}
\description{
The function extracts some important statistics from the model,
such as variance component, standard error, etc
and returns a data frame containing some important statistics.
}
