% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{filterDat}
\alias{filterDat}
\title{Filter the raw data}
\usage{
filterDat(dat, control = list(), verbose = TRUE)
}
\arguments{
\item{dat}{A data.frame format, The column name must contain 'Name','Range','Pass','Location','Field','Set','Type'.}

\item{control}{A list format, Extract the Trait, Locatiton, Field, and Type provided in the control list; Discard
outliers beyond n times (FilterStd) standard deviation.}

\item{verbose}{Whether to display the output information.}
}
\value{
Returns a list contains: A data frame containing columns for Name, Range, Pass, Location, Field, Type, Set,
trait,
  Qualified(Data conformity); A grouped data frame containnig columns for Location, Field, Type, Number(The number
  of samples in the group), unique_Range(Line number), unique_Pass(Column number), Pass(|Range| x |Pass| ?= |Name|),
  QCRate(Effective data volume).
}
\description{
The function is used for preliminary filtering of raw data, including:
1. filtering out the outlier specified by the user;
2. Process phenotypic outliers according to different Location, Field and Type (filter out outliers beyond n times
standard deviation).
}
\examples{
data("TestPheno")
controlList <- list(Region   = "EM",
                    Trait    = list(name = c("YLD", "MST", "PHT", "EHT"),
                                    max = c(10000, 10000, 10000, 10000),
                                    min = c(0, 0, 0, 0)),
                    Location = c("6HLJSC", "6JLDF", "6JLDH", "6JLGZL", "7LNTL", "7NMTL"),
                    Field    = c(1:8),
                    Type     = c("P1", "P3", "PLate"),
                    FilterStd = 3)
filterRes <- filterDat(dat = phenos[[1]], control = controlList)
filterRes$summary
}
