% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPbreedingValues_Estimation.R
\name{GCA_SCA_Estimation}
\alias{GCA_SCA_Estimation}
\title{Establishing multiple mixed linear models}
\usage{
GCA_SCA_Estimation(bluehb, ainv, ped)
}
\arguments{
\item{bluehb}{Data frame with blue value.}

\item{ainv}{Inverse matrix of kinship.}

\item{ped}{Genealogical file of the sample.}
}
\value{
A list with model results.
}
\description{
This function generates additive effect, non-additive effect,
generalized heritability and so on according to the blue value,
pedigree file and inverse matrix of genetic relationship.
}
