% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_arrange.R
\name{Hybrid_Observations}
\alias{Hybrid_Observations}
\title{Title}
\usage{
Hybrid_Observations(parent)
}
\arguments{
\item{parent}{a data.frame of parents}
}
\value{
result a data.frame of number of hybrid observations
}
\description{
Title
}
