% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mgxeEstimation}
\alias{mgxeEstimation}
\title{Multipoint (Location) joint analysis, BLUP EBV}
\usage{
mgxeEstimation(
  datdf,
  trait,
  ped = NULL,
  ainv = NULL,
  loc.grp.by = NULL,
  verbose = TRUE
)
}
\arguments{
\item{datdf}{A data frame format, phenotype data or BLUE data, The column needs to contain 'Name','Location','Field','Type', 'Trait'.}

\item{trait}{Target traits to be analyzed.}

\item{ped}{Pedigree file with three columns (ID, FGenoId, MGenoID).}

\item{ainv}{Inverse matrix of the genealogical matrix generated by asreml::ainverse(ped).}

\item{loc.grp.by}{Group locations.}

\item{verbose}{Whether to display the output information.}
}
\value{
Output the EBV values of parents in different environments.
}
\description{
This function uses multiple locations to estimate the EBV values of the F_parent and the M_parent.
}
\examples{
data('TestPheno')
data('TestAinv')
data('TestPed')
data("TestBlueList")
controlList <- list(Region   = "EM",
                    Trait    = list(name = c("YLD", "MST", "PHT", "EHT"),
                    max = c(10000, 10000, 10000, 10000),
                    min = c(0, 0, 0, 0)),
                    Location = c("6HLJSC", "6JLDF", "6JLDH", "6JLGZL", "7LNTL", "7NMTL"),
                    Field    = c(1:8),
                    Type     = c("P1", "P3", "PLate"),
                    FilterStd = 3)
filterRes <- filterDat(dat = phdemo[[1]], control = controlList)
grp.by  <- list(Spring=c("6JLGZL","7NMTL","7LNTL"),
                Summer=c("6HLJSC", "6JLDF", "6JLDH"))
testGxE <- mgxeEstimation(datdf = filterRes$filterData,
                          trait = "PHT",
                          ped   = ped,
                          ainv  = ainv,
                          loc.grp.by = grp.by)
plot(testGxE$mgxe$F.EBV.Spring$EBV, testGxE$mgxe$F.EBV.Summer$EBV)
plot(testGxE$mgxe$M.EBV.Spring$EBV, testGxE$mgxe$M.EBV.Summer$EBV)
}
