% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueEstimation1.0.R
\name{missingAdding}
\alias{missingAdding}
\title{Adding Missing positions for spatial analysis}
\usage{
missingAdding(w)
}
\arguments{
\item{w}{A data frame}
}
\value{
A data frame with missing value added
}
\description{
Adding Missing positions for spatial analysis
}
\examples{
\donttest{
 df <- data.frame(read.table(textConnection(
   '     Range Pass Set Rep     Name  YLD
  1    1  24   2  ID1   901
  1    3  24   2  ID2   852
  1    4  24   2  ID3   753
  2    1  23   2  ID4   765
  2    2  23   2  ID5   964
  2    3  23   2  ID6   963
  2    4  23   2  ID7   737
  3    1  22   2  ID8   803
  3    2  22   2  ID9   963
  3    3  22   2 ID10   769
  3    4  22   2 ID11   852
  4    1  21   2 ID12   778
  4    2  21   2 ID13   1021
  4    3  21   2 ID14   667
  4    4  21   2 ID15   852'
  ),header=T,sep=''))

print(table(df$Range,df$Pass))
dat <- missingAdding(w=df)
dat <- dat[order(dat$Range,dat$Pass),]
print(table(dat$Range,dat$Pass))
dat
}
}
