% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{mulTraitEstimation}
\alias{mulTraitEstimation}
\title{Multi-trait phenotypic correlation}
\usage{
mulTraitEstimation(datdf, trait.vec, ped = NULL, ainv = NULL, verbose = TRUE)
}
\arguments{
\item{datdf}{A data frame format, phenotype data or BLUE data, The column needs to contain 'Name','Location','Field','Type', 'Trait'.}

\item{trait.vec}{The set of traits to be analyzed.}

\item{ped}{Pedigree file with three columns (ID, FGenoId, MGenoID).}

\item{ainv}{Inverse matrix of the genealogical matrix generated by asreml::ainverse(ped).}

\item{verbose}{Whether to display the output information.}
}
\value{
Phenotypic correlation matrix of different traits.
}
\description{
Computational phenotypic correlation.
}
\examples{
data("TestAinv")
data("TestPed")
data("TestBlueList")
filter.res <- dataOutliers.for(data   = phdemo$SpML,
                               traits = c('MST',"YLD","PHT","EHT"),
                               model  = '3SD')
testCor <- mulTraitEstimation(datdf = filter.res$result_df,
                              trait.vec = c("YLD", "MST.x", "MST.y", "PHT", "EHT"),
                              ainv = ainv, ped = ped)
testCor
}
