% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_checking.R
\name{outliersRemoval}
\alias{outliersRemoval}
\title{Identify and process outliers in the data for a single trait}
\usage{
outliersRemoval(data, trait, model)
}
\arguments{
\item{data}{The input data as a data.frame.}

\item{trait}{The name of the trait to be processed.}

\item{model}{The method to be used for identifying outliers, either '3SD' or 'boxplot'.}
}
\value{
A list containing processed data for the given trait.
}
\description{
This function identifies and processes outliers in the data for a single trait.
}
\examples{
w <- phdemo$SpML
mst.Outliers<-dataOutliers_Identifying(data=w,trait ="MST",model='3SD')
}
