% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ped}
\alias{ped}
\title{Pedigree data.}
\format{
A data frame with 2820 rows and 3 variables:
\describe{
  \item{ID}{Individual ID}
  \item{FGenoId}{Male parent ID}
  \item{MGenoId}{Female parent ID}
}
}
\usage{
ped
}
\description{
Original pedigree data corresponding to test pheno data.
}
\keyword{datasets}
