% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phdemo}
\alias{phdemo}
\title{Phenotypic list demo data.}
\format{
A Large matrix with 31689 elements(By asreml::ainverse):
}
\usage{
phdemo
}
\description{
Contains the test data of different periods, regions and communities.
}
\keyword{datasets}
