% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLUPbreedingValues_Estimation.R
\name{programRun}
\alias{programRun}
\title{Calculating the statistics of all characters to be processed}
\usage{
programRun(bluelist, ainv, ped, traits)
}
\arguments{
\item{bluelist}{List with blue value.}

\item{ainv}{Inverse matrix of kinship.}

\item{ped}{Genealogical file of the sample.}

\item{traits}{All characters to be treated.}
}
\value{
A list containing statistics of all traits.
}
\description{
This function processes each trait in turn, returns their respective
statistical values and summarizes them in a list.
}
\examples{
\donttest{
library(biosim)
library(matrixcalc)
library(asreml)
library(MCMCglmm)
names(ped)[names(ped)=='FGenoID'] <- 'DAM'
names(ped)[names(ped)=='MGenoID'] <- 'SIRE'
print(head(ped))
bluedat <- lapply(bluedat, function(x) x$blue)
traits <- c('MST', 'TWT', 'YLD', 'PHT', 'EHT')
ebv <- programRun(bluedat[['SpML']], ainv, ped, traits)
print(names(ebv))
}

}
