% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blueEstimation1.0.R
\name{replicationCheck}
\alias{replicationCheck}
\title{the ratio of the number of unique values occurring more than once to the total number of unique values in the "Name" column}
\usage{
replicationCheck(df)
}
\arguments{
\item{df}{a data.frame}
}
\value{
a number
}
\description{
the ratio of the number of unique values occurring more than once to the total number of unique values in the "Name" column
}
\examples{
\donttest{
 df <- data.frame(read.table(textConnection(
   '   Name Range Pass Set Rep YLD
  1    1  24   2  ID1   901
  2    3  24   2  ID2   852
  2    4  24   2  ID3   753
  2    1  23   2  ID4   765
  2    2  23   2  ID5   964
  2    3  23   2  ID6   963
  2    4  23   2  ID7   737
  3    1  22   2  ID8   803
  3    2  22   2  ID9   963
  3    3  22   2 ID10   769
  3    4  22   2 ID11   852
  4    1  21   2 ID12   778
  4    2  21   2 ID13   1021
  4    3  21   2 ID14   667
  4    4  21   2 ID15   852'
  ),header=T,sep=''))

print(table(df$Name))
dat <- replicationCheck(df)
dat
}
}
